refactor(gui): split app.js into ES modules
Pure code motion: the 3,770-line frontend entry becomes eleven modules —
dom (element builder + formatters), markdown (parser/renderer/streaming
view), transcript (flow + row builders), pane (image/diff panes and the
pane plumbing), context (context panel + git polling), subagents, attach,
palette, composer, settings — with app.js keeping the state object, the
stream callbacks, task selection, and boot. The new files are embedded
and routed in the GUI server alongside the existing three.
